requirements-insights.txt
# used for metrics
opencensus-ext-azure>=1.1.8